fix compile warning with gcc 4.1

+++ starka/src/lib/starling_common/test/depth_buffer_test.cpp
@@ -40,7 +40,7 @@ get_db_test_pattern() {
 BOOST_AUTO_TEST_CASE( test_depth_buffer_val ) {
 
     depth_buffer db(get_db_test_pattern());
-    BOOST_CHECK_EQUAL(db.val(109),9);
+    BOOST_CHECK_EQUAL(static_cast<int>(db.val(109)),9);
 }
 
 
@@ -48,7 +48,7 @@ BOOST_AUTO_TEST_CASE( test_depth_buffer_clear ) {
 
     depth_buffer db(get_db_test_pattern());
     db.clear_pos(109);
-    BOOST_CHECK_EQUAL(db.val(109),0);
+    BOOST_CHECK_EQUAL(static_cast<int>(db.val(109)),0);
 }
